Added GDPR compliance stuff
- There is an opt-in checkbox for remembering your last search.
- A tooltip under the checkbox explains the GDPR stuff

With the checkbox unchecked, the site uses session cookie for
normal operation. It is the browser's responsibility to delete
session cookies after the session is closed.

In testing, it was found that all browsers don't "close" the
session after closing the tab/window. There is nothing I can do
about this.

Regardless, the browser's configuration, the cookie information is
never sent, unless the user chooses to send the information.

The two ways this can be done are:
- The user writes a name of a city and clicks search.
- The user checks "Remember Me" and the user opens or refreshes
  the page.