Fixed a severe state issue
**State issue**

weather.om allowed a user to change the server state, such that
the state change was applied to other users. (model selection)

The library was fixed and the code was adapted to use the fixed
libary. Now model is defined for every request, such that the
API is truly stateless - as it should be.

**Added more testin**

Added unit testing for input sanitization functions

**UX**

The browser now stores the model in a cookie and if allowed, will
store the cookie accross sessions.